[multi]
- move value into its own module
- remove _ from match clause
